# shared/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VRA-Trust Evidence API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/diagnosis": {
      "get": {
        "summary": "Diagnosis",
        "operationId": "diagnosis_api_projects__project_id__diagnosis_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or origin rejected"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/projects/{project_id}/materials": {
      "get": {
        "summary": "Materials",
        "operationId": "materials_api_projects__project_id__materials_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or origin rejected"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "post": {
        "summary": "Create Material",
        "operationId": "create_material_api_projects__project_id__materials_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialCard"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or origin rejected"
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ]
      }
    },
    "/api/projects/{project_id}/materials/{material_id}": {
      "put": {
        "summary": "Update Material",
        "operationId": "update_material_api_projects__project_id__materials__material_id__put",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "material_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Material Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or origin rejected"
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ]
      }
    },
    "/api/projects/{project_id}/materials/{material_id}/history": {
      "get": {
        "summary": "Material History",
        "operationId": "material_history_api_projects__project_id__materials__material_id__history_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "material_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Material Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or origin rejected"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/projects/{project_id}/material-targets/{run_id}": {
      "get": {
        "summary": "Material Targets",
        "operationId": "material_targets_api_projects__project_id__material_targets__run_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or origin rejected"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/projects/{project_id}/studies": {
      "post": {
        "summary": "Start Study",
        "operationId": "start_study_api_projects__project_id__studies_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or origin rejected"
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ]
      },
      "get": {
        "summary": "Studies",
        "operationId": "studies_api_projects__project_id__studies_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or origin rejected"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/projects/{project_id}/studies/{study_id}": {
      "get": {
        "summary": "Study",
        "operationId": "study_api_projects__project_id__studies__study_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Study Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or origin rejected"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/projects/{project_id}/studies/{study_id}/reevaluate": {
      "post": {
        "summary": "Reevaluate",
        "operationId": "reevaluate_api_projects__project_id__studies__study_id__reevaluate_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Study Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or origin rejected"
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ]
      }
    },
    "/api/projects/{project_id}/studies/{study_id}/report.json": {
      "get": {
        "summary": "Report Json",
        "operationId": "report_json_api_projects__project_id__studies__study_id__report_json_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Study Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or origin rejected"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/projects/{project_id}/studies/{study_id}/report.html": {
      "get": {
        "summary": "Report Html",
        "operationId": "report_html_api_projects__project_id__studies__study_id__report_html_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Study Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or origin rejected"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/projects/{project_id}/studies/{study_id}/report.pdf": {
      "get": {
        "summary": "Report Pdf",
        "operationId": "report_pdf_api_projects__project_id__studies__study_id__report_pdf_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Study Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or origin rejected"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/projects/{project_id}/studies/{study_id}/artifacts": {
      "get": {
        "summary": "Artifacts",
        "operationId": "artifacts_api_projects__project_id__studies__study_id__artifacts_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Study Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or origin rejected"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaterialCard": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Name"
          },
          "batch": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch"
          },
          "contains_rare_earth": {
            "type": "boolean",
            "title": "Contains Rare Earth",
            "default": false
          },
          "conductivity_w_mk": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Conductivity W Mk"
          },
          "density_kg_m3": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 30000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Density Kg M3"
          },
          "specific_heat_j_kgk": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 20000.0,
                "minimum": 100.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Specific Heat J Kgk"
          },
          "thermal_evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thermal Evidence Id"
          },
          "sample_evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Evidence Id"
          },
          "control_evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Control Evidence Id"
          },
          "aging_evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aging Evidence Id"
          },
          "test_conditions": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Conditions"
          },
          "applicability": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicability"
          },
          "source_locator": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Locator"
          },
          "review_state": {
            "type": "string",
            "enum": [
              "pending",
              "confirmed"
            ],
            "title": "Review State",
            "default": "pending"
          },
          "responsible_person": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Responsible Person"
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "responsible_person"
        ],
        "title": "MaterialCard"
      },
      "MaterialUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Name"
          },
          "batch": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch"
          },
          "contains_rare_earth": {
            "type": "boolean",
            "title": "Contains Rare Earth",
            "default": false
          },
          "conductivity_w_mk": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Conductivity W Mk"
          },
          "density_kg_m3": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 30000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Density Kg M3"
          },
          "specific_heat_j_kgk": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 20000.0,
                "minimum": 100.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Specific Heat J Kgk"
          },
          "thermal_evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thermal Evidence Id"
          },
          "sample_evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Evidence Id"
          },
          "control_evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Control Evidence Id"
          },
          "aging_evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aging Evidence Id"
          },
          "test_conditions": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Conditions"
          },
          "applicability": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicability"
          },
          "source_locator": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Locator"
          },
          "review_state": {
            "type": "string",
            "enum": [
              "pending",
              "confirmed"
            ],
            "title": "Review State",
            "default": "pending"
          },
          "responsible_person": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Responsible Person"
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "responsible_person",
          "expected_revision"
        ],
        "title": "MaterialUpdate"
      },
...
      "StudyCreate": {
        "properties": {
          "source_run_id": {
            "type": "string",
            "title": "Source Run Id"
          },
          "target_material": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Target Material"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/Option"
            },
            "type": "array",
            "maxItems": 12,
            "minItems": 1,
            "title": "Options"
          },
          "objectives": {
            "items": {
              "type": "string",
              "enum": [
                "energy_kwh",
                "cost_cny",
                "cooling_unmet_h",
                "carbon_kg"
              ]
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 1,
            "title": "Objectives"
          },
          "constraints": {
            "$ref": "#/components/schemas/Constraints"
          },
          "budget": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Budget"
          },
          "basis": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 5,
            "title": "Basis"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_run_id",
          "target_material",
          "options",
          "budget",
          "basis"
        ],
        "title": "StudyCreate"
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "vra_session"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    }
  }
}